fix(acp): resume provider before fork response
